Add Grid.Ratio mode

+++ src/fe_present.hpp
@@ -230,7 +230,7 @@ protected:
 	const char *get_layout_font_name() const;
 	bool get_preserve_aspect_ratio();
 	bool get_layout_crop();
-	sf::Vector2i get_surface_texture_size( FePresentableParent &p, int w, int h ) const;
+	sf::Vector2i get_surface_texture_size( FePresentableParent &p, float w, float h ) const;
 
 	void set_selection_index( int );
 	void set_layout_width( float );
